requirements.txt
anthropic>=0.40.0
pyyaml>=6.0
schedule>=1.2.0
beautifulsoup4>=4.12.0
httpx>=0.25.0
jinja2>=3.1.0
python-dateutil>=2.8.0
